backend/handout/serializer/categorySerializer.py
from rest_framework import serializers
from handout.models import Category

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'name', 'slug', 'parent']