backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
django-filter==24.2
django-jalali==6.2.1
djangorestframework==3.15.2
Markdown==3.6
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
